refactor: cleanup `ChunkCache` implementations
- Single definitions of each chunk cache
  - Makes `cbindgen` happy for `zarrs_ffi`
  - No false gates in docs
- Platform implementation details fully private
- Remove global macros
- Cleanup WASM warnings
