+++ zarrs/src/array/array_sync_sharded_readable_ext.rs
@@ -1,7 +1,10 @@
 use std::{collections::HashMap, sync::Arc};
 
 use crate::iter_concurrent_limit;
+
+#[cfg(not(target_arch = "wasm32"))]
 use rayon::iter::{IntoParallelIterator, ParallelIterator};
+
 use unsafe_cell_slice::UnsafeCellSlice;
 use zarrs_metadata::ConfigurationSerialize;
 use zarrs_metadata_ext::codec::sharding::ShardingCodecConfiguration;
