+++ zarrs/src/array/chunk_cache.rs
@@ -1,7 +1,9 @@
 use std::sync::Arc;
 
 use crate::iter_concurrent_limit;
+#[cfg(not(target_arch = "wasm32"))]
 use rayon::iter::{IntoParallelIterator, ParallelIterator};
+
 use unsafe_cell_slice::UnsafeCellSlice;
 use zarrs_metadata::DataTypeSize;
 use zarrs_storage::ReadableStorageTraits;
@@ -20,7 +22,7 @@ use crate::{
 use super::{codec::CodecOptions, ArrayBytes, ArrayError, RawBytes};
 
 pub(crate) mod chunk_cache_lru;
-pub(crate) mod chunk_cache_lru_macros;
+// pub(crate) mod chunk_cache_lru_macros;
 
 /// The chunk type of an encoded chunk cache.
 pub type ChunkCacheTypeEncoded = Option<Arc<RawBytes<'static>>>;
@@ -31,7 +33,7 @@ pub type ChunkCacheTypeDecoded = Arc<ArrayBytes<'static>>;
 /// The chunk type of a partial decoder chunk cache.
 pub type ChunkCacheTypePartialDecoder = Arc<dyn ArrayPartialDecoderTraits>;
 
-/// A chunk type ([`ChunkCacheTypeEncoded`], [`ChunkCacheTypeDecoded`], or [`ChunkCacheTypePartialDecoder`]).
+/// A chunk cache type ([`ChunkCacheTypeEncoded`], [`ChunkCacheTypeDecoded`], or [`ChunkCacheTypePartialDecoder`]).
 pub trait ChunkCacheType: MaybeSend + MaybeSync + Clone + 'static {
     /// The size of the chunk in bytes.
     fn size(&self) -> usize;
